requirements.txt
numpy==1.24.4
Pillow==10.1.0
tensorflow-cpu==2.15.0
